feat: set of improvements for monitoring
